requirements.txt
Django>=4.1,<4.2
wagtail==4.0.4
django-widget-tweaks
django-environ
djangorestframework
django-cors-headers
gunicorn==20.1.0
whitenoise==6.2.0

psycopg2==2.9.3

redis==4.3.4
celery==5.2.7

flake8==5.0.4
black==22.8.0

django-crispy-forms==1.14.0
django-phonenumber-field==7.0.0
phonenumbers==8.12.57